feat: extract the outbound staging verdict and delta computation into the pure engine (#288)
The third slice of the #288 outbound unbraiding (plan Phase 1b). The
verdict behind the three CreateOrUpdatePendingExport* entry points
(Object Type conflict, provisioning gate, recall guard, pending
provisioning reuse or irrelevance, update) moves into
SyncEngine.DecideOutboundStaging; the in-memory Pending Export merge
(#1199 whole-attribute supersede included) becomes
SyncEngine.MergeAttributeChangesIntoPendingExport; and the outbound
delta computation (CreateAttributeValueChanges and its comparison and
dictionary helpers) moves wholesale into the engine as
ComputeAttributeValueChanges, taking the expression evaluator as a
parameter in the engine's established style. ExportEvaluationServer
keeps orchestration (export matching, CSO creation, persistence) and
thin delegating wrappers so every existing caller and characterisation
test exercises the same semantics through the same entry points.

18 new red-first engine tests pin the verdict and the merge; the
existing OutboundSync characterisation suites pin the moved
computation unchanged.
